auth_project/requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.1
pika==1.1.0
pika-stubs==0.1.3
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.4.1
uuid==1.30
